Dataloader/requirements.txt
apache-airflow-providers-google==8.0.0
apache-beam[gcp]==2.65.0
flask==2.2.2
google-cloud-bigquery==3.10.0
google-cloud-datacatalog==3.27.1
google-cloud-logging==3.3.0
google-cloud-pubsub==2.16.4
google-cloud-storage==2.7.0
pytest==7.2.0
requests==2.28.2
